Add files via upload
The Python script is a port scanner that allows the user to scan a target IP address for open ports within a specified range. It uses the Scapy library to construct and send TCP SYN packets to the target IP's ports. The response from the target determines whether the port is open or not.